Sat, Mar 25, 2023, 11:33 AM +01:00

+++ filter-less/helpers.c
@@ -76,39 +76,39 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
         {
             if (h == 0 && w == 0)
             {
-                copy[h][w].rgbtRed = round(image[h][w].rgbtRed + image[h][w + 1].rgbtRed
-                                           + image[h + 1][w + 1].rgbtRed + image[h + 1][w].rgbtRed) / 4;
-                copy[h][w].rgbtGreen = round(image[h][w].rgbtGreen + image[h][w + 1].rgbtGreen
-                                             + image[h + 1][w + 1].rgbtGreen + image[h + 1][w].rgbtGreen) / 4;
-                copy[h][w].rgbtBlue = round(image[h][w].rgbtBlue + image[h][w + 1].rgbtBlue
-                                            + image[h + 1][w + 1].rgbtBlue + image[h + 1][w].rgbtBlue) / 4;
+                copy[h][w].rgbtRed = (round((double)(image[h][w].rgbtRed + image[h][w + 1].rgbtRed
+                                           + image[h + 1][w + 1].rgbtRed + image[h + 1][w].rgbtRed) / 4));
+                copy[h][w].rgbtGreen = (round((double)(image[h][w].rgbtGreen + image[h][w + 1].rgbtGreen
+                                             + image[h + 1][w + 1].rgbtGreen + image[h + 1][w].rgbtGreen) / 4));
+                copy[h][w].rgbtBlue = (round((double)(image[h][w].rgbtBlue + image[h][w + 1].rgbtBlue
+                                            + image[h + 1][w + 1].rgbtBlue + image[h + 1][w].rgbtBlue) / 4));
             }
             else if (h == 0 && w == width - 1)
             {
-                copy[h][w].rgbtRed = round(image[h][w].rgbtRed + image[h + 1][w].rgbtRed
-                                           + image[h + 1][w - 1].rgbtRed + image[h][w - 1].rgbtRed) / 4;
-                copy[h][w].rgbtGreen = round(image[h][w].rgbtGreen + image[h + 1][w].rgbtGreen
-                                             + image[h + 1][w - 1].rgbtGreen + image[h][w - 1].rgbtGreen) / 4;
-                copy[h][w].rgbtBlue = round(image[h][w].rgbtBlue + image[h + 1][w].rgbtBlue
-                                            + image[h + 1][w - 1].rgbtBlue + image[h][w - 1].rgbtBlue) / 4;
+                copy[h][w].rgbtRed = (round((double)(image[h][w].rgbtRed + image[h + 1][w].rgbtRed
+                                           + image[h + 1][w - 1].rgbtRed + image[h][w - 1].rgbtRed) / 4));
+                copy[h][w].rgbtGreen = (round((double)(image[h][w].rgbtGreen + image[h + 1][w].rgbtGreen
+                                             + image[h + 1][w - 1].rgbtGreen + image[h][w - 1].rgbtGreen) / 4));
+                copy[h][w].rgbtBlue = (round((double)(image[h][w].rgbtBlue + image[h + 1][w].rgbtBlue
+                                            + image[h + 1][w - 1].rgbtBlue + image[h][w - 1].rgbtBlue) / 4));
             }
             else if (h == height - 1 && w == 0)
             {
-                copy[h][w].rgbtRed = round(image[h][w].rgbtRed + image[h - 1][w].rgbtRed
-                                           + image[h - 1][w + 1].rgbtRed + image[h][w + 1].rgbtRed) / 4;
-                copy[h][w].rgbtGreen = round(image[h][w].rgbtGreen + image[h - 1][w].rgbtGreen
-                                             + image[h - 1][w + 1].rgbtGreen + image[h][w + 1].rgbtGreen) / 4;
-                copy[h][w].rgbtBlue = round(image[h][w].rgbtBlue + image[h - 1][w].rgbtBlue
-                                            + image[h - 1][w + 1].rgbtBlue + image[h][w + 1].rgbtBlue) / 4;
+                copy[h][w].rgbtRed = (round((double)(image[h][w].rgbtRed + image[h - 1][w].rgbtRed
+                                           + image[h - 1][w + 1].rgbtRed + image[h][w + 1].rgbtRed) / 4));
+                copy[h][w].rgbtGreen = (round((double)(image[h][w].rgbtGreen + image[h - 1][w].rgbtGreen
+                                             + image[h - 1][w + 1].rgbtGreen + image[h][w + 1].rgbtGreen) / 4));
+                copy[h][w].rgbtBlue = (round((double)(image[h][w].rgbtBlue + image[h - 1][w].rgbtBlue
+                                            + image[h - 1][w + 1].rgbtBlue + image[h][w + 1].rgbtBlue) / 4));
             }
             else if (h == height - 1 && w == width - 1)
             {
-                copy[h][w].rgbtRed = round((double)image[h][w].rgbtRed + image[h - 1][w].rgbtRed
+                copy[h][w].rgbtRed = (round((double)(image[h][w].rgbtRed + image[h - 1][w].rgbtRed
                                            + image[h][w - 1].rgbtRed + image[h - 1][w - 1].rgbtRed) / 4));
-                copy[h][w].rgbtGreen = round((double)image[h][w].rgbtGreen + image[h - 1][w].rgbtGreen
-                                             + image[h][w - 1].rgbtGreen + image[h - 1][w - 1].rgbtGreen) / 4;
-                copy[h][w].rgbtBlue = round((double)image[h][w].rgbtBlue + image[h - 1][w].rgbtBlue
-                                            + image[h][w - 1].rgbtBlue + image[h - 1][w - 1].rgbtBlue) / 4;
+                copy[h][w].rgbtGreen = (round((double)(image[h][w].rgbtGreen + image[h - 1][w].rgbtGreen
+                                             + image[h][w - 1].rgbtGreen + image[h - 1][w - 1].rgbtGreen) / 4));
+                copy[h][w].rgbtBlue = (round((double)(image[h][w].rgbtBlue + image[h - 1][w].rgbtBlue
+                                            + image[h][w - 1].rgbtBlue + image[h - 1][w - 1].rgbtBlue) / 4));
             }
             else if (h == 0 && 0 < w && w < (width - 1))
             {
